local build of OCaml is not supported anymore
because local compilation of OCaml depends on camlp4 and num present in OCaml distribution, but
camlp4 was removed in 4.02 and we need at least 4.02 for Bytes to support newest OCaml versions.
Supporting local build of multiple components is too much work, maybe we
can bring it back as opam bundle in the future.